+++ src/pharmpy/data/data_frame.py
@@ -234,22 +234,6 @@ class DataFrameAccessor:
             time_var = self._obj.groupby(by=self.id_label)[cov_labels].nunique().gt(1).any()
             return list(time_var.index[time_var])
 
-    @property
-    def doses(self):
-        """Return a series with all doses. Indexed with ID and TIME"""
-        try:
-            label = self.labels_by_type[ColumnType.DOSE]
-        except KeyError:
-            raise DatasetError('Could not identify dosing rows in dataset')
-
-        label = label[0]
-        idcol = self.id_label
-        idvcol = self.idv_label
-        df = self._obj.query(f'{label} != 0')
-        df = df[[idcol, idvcol, label]]
-        df.set_index([idcol, idvcol], inplace=True)
-        return df.squeeze()
-
     def add_doseid(self):
         """Add a column DOSEID with id of each dose period starting from 1"""
         try: