+++ src/pharmpy/modeling/__init__.py
@@ -28,6 +28,7 @@ from pharmpy.modeling.covariate_effect import add_covariate_effect
 from pharmpy.modeling.data import (
     get_baselines,
     get_covariate_baselines,
+    get_doses,
     get_ids,
     get_mdv,
     get_number_of_individuals,
@@ -129,6 +130,7 @@ __all__ = [
     'fix_parameters_to',
     'get_baselines',
     'get_covariate_baselines',
+    'get_doses',
     'get_ids',
     'get_number_of_individuals',
     'get_number_of_observations',